+++ frontend/src/container/ListOfDashboard/components/TableComponents/Date.tsx
@@ -2,16 +2,16 @@ import { Typography } from 'antd';
 import convertDateToAmAndPm from 'lib/convertDateToAmAndPm';
 import getFormattedDate from 'lib/getFormatedDate';
 
-import { Data } from '..';
+import { Data } from '../..';
 
-function Created(createdBy: Data['createdBy']): JSX.Element {
-	const time = new Date(createdBy);
+function DateComponent(lastUpdatedTime: Data['lastUpdatedTime']): JSX.Element {
+	const time = new Date(lastUpdatedTime);
 
 	const date = getFormattedDate(time);
 
 	const timeString = `${date} ${convertDateToAmAndPm(time)}`;
 
-	return <Typography>{`${timeString}`}</Typography>;
+	return <Typography>{timeString}</Typography>;
 }
 
-export default Created;
+export default DateComponent;
